+++ shop.html
@@ -154,7 +154,7 @@
                 </div>
                 <br>
                 <button id="filter-search">Search</button>
-                <script src="filter.js"></script>
+                <script type="module" src="filter.js"></script>
                 
             
                 <!-- filtering search menu-->
